Split registry-related code in a separate file
